DAOS-12383 common: Persist DAV heap statistics - counter related to currently allocated memory is now persisted   at the end of every transaction. - Removed check for sysdb in umem_pobj_open() - Fixed integer overflow in off2lba() wich was inturn causing   failures while rehydrating large vos files.
Required-githooks: true

+++ src/bio/bio_wal.c
@@ -663,7 +663,7 @@ fill_trans_blks(struct bio_meta_context *mc, struct bio_sglist *bsgl, struct ume
 static inline uint64_t
 off2lba(struct wal_super_info *si, unsigned int blk_off)
 {
-	return (blk_off + WAL_HDR_BLKS) * si->si_header.wh_blk_bytes;
+	return (uint64_t)(blk_off + WAL_HDR_BLKS) * si->si_header.wh_blk_bytes;
 }
 
 struct wal_tx_desc {
